refactor global utils
