+++ assembler/assembler.hpp
@@ -7,7 +7,8 @@
 
 #include "token.hpp"
 #include "error.hpp"
-#include "utils.hpp"
+#include "asm_utils.hpp"
+#include "../common/constants.hpp"
 
 Instruction assemble(std::vector<Token> tokens, Error error);
