+++ build.sh
@@ -2,18 +2,19 @@ BUILD=build
 
 ASSEMBLER=$(find ./assembler -type f -name "*.cpp")
 EMULATOR=$(find ./emulator -type f -name "*.cpp")
+UTILS=$(find ./common -type f -name "*.cpp")
 
 assemble() {
 	mkdir -p $BUILD/assembler
 
-	g++ $ASSEMBLER -o $BUILD/assembler/main
+	g++ $UTILS $ASSEMBLER -o $BUILD/assembler/main
 	./$BUILD/assembler/main
 }
 
 emulate() {
 	mkdir -p $BUILD/emulator
 
-	g++ $EMULATOR -o $BUILD/emulator/main
+	g++ $UTILS $EMULATOR -o $BUILD/emulator/main
 	./$BUILD/emulator/main
 }
